requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.8.3
aiosignal==1.3.2
attrs==24.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.2
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
Flask==3.1.0
frozenlist==1.5.0
google-api-core==2.24.2
google-auth==2.40.2
google-cloud-vision==3.10.1
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
MouseInfo==0.1.3
multidict==6.1.0
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==11.0.0
propcache==0.2.1
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
PyJWT==2.10.1
PyMsgBox==1.0.9
PyPDF2==3.0.1
pypdfium2==4.30.0
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytweening==1.2.0
pywhatkit==5.4
reportlab==4.2.5
requests==2.32.3
rsa==4.9.1
schedule==1.2.2
soupsieve==2.6
twilio==9.4.3
urllib3==2.2.3
Werkzeug==3.1.3
wikipedia==1.4.0
yarl==1.18.3
